tests/requirements.txt
-r ../requirements.txt
pytest-cov
black
pylint
cfn-lint
moto